🗺️ Atlas: Extract tests from core::version to resolve Blob anti-pattern
